test(ambrosian): move the season stamp and buildContextFor() into the shared real-year harness
Two related consistency fixes plus the message wording they make necessary.

1. Harness/production divergence. `AmbrosianRealYearHarnessTrait` never called
   `stampAmbrosianSeasonOnSanctorale()`, which production DOES call
   (`CalendarHandler::calculateAmbrosianCalendar()`) and whose own docblock says
   MUST run before precedence resolution. One of five consumers patched around
   it inline; the other four resolved with `liturgical_season = null` on every
   sanctorale event, which silently disables the resolver's and the rank
   classifier's season-gated branches — already enough, once, to make a test's
   red and green indistinguishable. The call now lives in the trait and the
   inline patch is gone.

2. `buildContextFor()` was duplicated verbatim in four test classes, all of
   which already use the trait. Moved into the trait; the four copies and their
   now-unused imports are deleted.

3. `AmbrosianPrecedenceResolver`'s suppression message said "suppressed by the
   higher-ranking X". Under the composite key that is inaccurate on exactly the
   days the tiebreak exists for: the two events share Tabella rank 10 and only
   the tiebreak separates them. It now says "higher-precedence", and
   `AmbrosianRealYearBvmMemorialPrecedenceTest` pins the full literal message
   rather than merely the suppression outcome.
